fix(share): 화면 개선, 컴포넌트 분리, 미지원 sns 보류
